Apply code review suggestions(move the `RemoteControl.API.definition` call to server app)
