+++ apps/server/lib/lexical/server/provider/handlers/go_to_definition.ex
@@ -1,12 +1,12 @@
 defmodule Lexical.Server.Provider.Handlers.GoToDefinition do
   alias Lexical.Protocol.Requests.GoToDefinition
   alias Lexical.Protocol.Responses
-  alias Lexical.RemoteControl.CodeIntelligence.Entity
+  alias Lexical.RemoteControl
 
   require Logger
 
   def handle(%GoToDefinition{} = request, env) do
-    case Entity.definition(env.project, request.document, request.position) do
+    case RemoteControl.Api.definition(env.project, request.document, request.position) do
       {:ok, native_location} ->
         {:reply, Responses.GoToDefinition.new(request.id, native_location)}
